Add viewport meta tag to index.html, and create responsive breakpoint for 500px mobile devices.
Style adjustments to make site match closer to provided design files.

Reset title-container's div and hr width to 100%;
Do the same for the .artist sections, disable crop on images,
automatically set height for images, and set width to reasonable amount.

Set artist's container to be column instead of row, and re-align-content
to the center instead of the left.